Cherry-picked f28e440e5435514b1eda19cbadfb12f4816f1290 Withot that I get flooded with: mpp_log: Assertion prep->ver_stride == (((prep->height)+(8)-1)&~((8)-1)) failed at h264e_vpu_update_hw_cfg:404 mpp_log: Assertion prep->ver_stride == (((prep->height)+(8)-1)&~((8)-1)) failed at h264e_vpu_update_hw_cfg:404 mpp_log: Assertion prep->ver_stride == (((prep->height)+(8)-1)&~((8)-1)) failed at h264e_vpu_update_hw_cfg:404 mpp_log: Assertion prep->ver_stride == (((prep->height)+(8)-1)&~((8)-1)) failed at h264e_vpu_update_hw_cfg:404 mpp_log: Assertion prep->ver_stride == (((prep->height)+(8)-1)&~((8)-1)) failed at h264e_vpu_update_hw_cfg:404 mpp_log: Assertion prep->ver_stride == (((prep->height)+(8)-1)&~((8)-1)) failed at h264e_vpu_update_hw_cfg:404
I don't know if that is the right patch though,
as the author of the bug report https://forum.radxa.com/t/video-encoding-on-rock-pi-4b/2046
didn't specify the commit (only roughly a date).

video: rockchip: vpu: fix the client type configure

The userspace would use the client type ioctl to make
a session work in secure mode, reuse the undefine
bits of the argument.

This patch would remove the secure mode setting from
the client type argument which was a mistake in a
previous commit, but without the secure mode applied,
nobody would know.

Change-Id: Ifaa976a0c9ab0c981b1f94f4978cd13a61adb0a7
Signed-off-by: Randy Li <randy.li@rock-chips.com>